Changed reload to restart when changing a zone.
When a zone is changed, in order to notify the slaves and delete any
negative cache BIND needs to be restarted. This could be avoided but
will require connecting to the servers and issuing either a restart in
the master when a register is cached or to the slaves to force updates.